backend/requirements.txt
# Core Django
Django==5.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.3

# Environment
python-decouple==3.8

# API Docs
drf-spectacular==0.28.0

# Static files
whitenoise==6.9.0

# Image handling
Pillow==12.0.0

# HTTP requests (for aggregation)
requests==2.32.5
feedparser==6.0.11

# Cloudinary (media storage)
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Production server
gunicorn==23.0.0

# Utilities
python-slugify==8.0.4
asgiref==3.11.0
sqlparse==0.5.3

# Phase 3 Intelligence (AI/Vector)
openai==1.54.3
pgvector==0.2.5
sentence-transformers==2.7.0
numpy==1.26.4
